chapter_001/src/main/java/ru/job4j/array/Check.java
package ru.job4j.array;

/**
 * @author Kosolapov Ilya (d_dexter@mail.ru)
 * @version $ID$
 * @since 0.1
 */
public class Check {

    /**
     * Checkout all elements equals among themselves.
     * @param data array.
     * @return check.
     */
    public boolean mono(boolean[] data) {
        boolean result = true;
        for (int i = 1; i < data.length; i++) {
            if (data[0] != data[i]) {
                result = false;
                break;
            }
        }
        return result;
    }
}